LayoutTests/fast/css-grid-layout/grid-self-baseline-two-dimensional-expected.txt


FAIL .grid 1 assert_equals:
<div class="grid fit-content itemsBaseline">
        <span class="firstRowFirstColumn verticalLR font12 extraTopPadding" data-offset-x="25" data-offset-y="10">A<br>A</span><span class="firstRowSecondColumn horizontalTB font16 extraBottomPadding" data-offset-x="183" data-offset-y="35">A<br>A</span><span class="firstRowThirdColumn verticalLR font24" data-offset-x="371" data-offset-y="10">A<br>A</span><span class="firstRowFourthColumn horizontalTB font32" data-offset-x="505" data-offset-y="10">A<br>A</span>
        <span class="secondRowFirstColumn horizontalTB font16" data-offset-x="17" data-offset-y="192">A<br>A</span><span class="secondRowSecondColumn verticalLR font24" data-offset-x="186" data-offset-y="175">A<br>A</span><span class="secondRowThirdColumn horizontalTB font12 extraTopPadding" data-offset-x="349" data-offset-y="175">A<br>A</span><span class="secondRowFourthColumn verticalLR font32" data-offset-x="505" data-offset-y="175">A<br>A</span>
        <span class="thirdRowFirstColumn verticalLR font32" data-offset-x="17" data-offset-y="293">A<br>A</span><span class="thirdRowSecondColumn horizontalTB font12 extraTopPadding" data-offset-x="183" data-offset-y="300">A<br>A</span><span class="thirdRowThirdColumn verticalLR font24 extraLeftPadding" data-offset-x="349" data-offset-y="293">A<br>A</span><span class="thirdRowFourthColumn horizontalTB font16 extraTopPadding" data-offset-x="505" data-offset-y="293">A<br>A</span>
        <span class="fourthRowFirstColumn horizontalTB font24" data-offset-x="17" data-offset-y="418">A<br>A</span><span class="fourthRowSecondColumn verticalLR font32" data-offset-x="183" data-offset-y="418">A<br>A</span><span class="fourthRowThirdColumn horizontalTB font16" data-offset-x="349" data-offset-y="430">A<br>A</span><span class="fourthRowFourthColumn verticalLR font12" data-offset-x="513" data-offset-y="418">A<br>A</span>
    </div>
offsetTop expected 293 but got 294
FAIL .grid 2 assert_equals:
<div class="grid fit-content itemsBaseline">
        <span class="firstRowFirstColumn verticalLR font12 extraLeftPadding" data-offset-x="17" data-offset-y="10">A<br>A</span><span class="firstRowSecondColumn horizontalTB font16 extraBottomPadding" data-offset-x="197" data-offset-y="35">A<br>A</span><span class="firstRowThirdColumn verticalLR font24" data-offset-x="299" data-offset-y="10">A<br>A</span><span class="firstRowFourthColumn horizontalTB font32" data-offset-x="433" data-offset-y="10">A<br>A</span>
        <span class="secondRowFirstColumn verticalLR font16 extraTopPadding" data-offset-x="38" data-offset-y="175">A<br>A</span><span class="secondRowSecondColumn horizontalTB font24" data-offset-x="197" data-offset-y="188">A<br>A</span><span class="secondRowThirdColumn verticalLR font12 extraTopPadding" data-offset-x="304" data-offset-y="175">A<br>A</span><span class="secondRowFourthColumn horizontalTB font32 extraLeftPadding" data-offset-x="433" data-offset-y="175">A<br>A</span>
        <span class="thirdRowFirstColumn verticalLR font32" data-offset-x="31" data-offset-y="340">A<br>A</span><span class="thirdRowSecondColumn horizontalTB font12 extraTopPadding" data-offset-x="197" data-offset-y="340">A<br>A</span><span class="thirdRowThirdColumn verticalLR font24" data-offset-x="299" data-offset-y="340">A<br>A</span><span class="thirdRowFourthColumn horizontalTB font16" data-offset-x="433" data-offset-y="357">A<br>A</span>
        <span class="fourthRowFirstColumn verticalLR font24" data-offset-x="34" data-offset-y="458">A<br>A</span><span class="fourthRowSecondColumn horizontalTB font32" data-offset-x="197" data-offset-y="458">A<br>A</span><span class="fourthRowThirdColumn verticalLR font16 extraBottomPadding" data-offset-x="303" data-offset-y="458">A<br>A</span><span class="fourthRowFourthColumn horizontalTB font12 extraTopPadding" data-offset-x="433" data-offset-y="466">A<br>A</span>
    </div>
offsetLeft expected 38 but got 37
PASS .grid 3
FAIL .grid 4 assert_equals:
<div class="grid fit-content itemsBaseline verticalRL">
        <span class="firstRowFirstColumn verticalLR font12" data-offset-x="328" data-offset-y="50">A</span><span class="firstRowSecondColumn horizontalTB font16 extraBottomPadding" data-offset-x="323" data-offset-y="208">A</span><span class="firstRowThirdColumn verticalLR font24" data-offset-x="323" data-offset-y="300">A</span><span class="firstRowFourthColumn horizontalTB font32" data-offset-x="323" data-offset-y="457">A</span>
        <span class="secondRowFirstColumn horizontalTB font16 extraBottomPadding" data-offset-x="253" data-offset-y="63">A</span><span class="secondRowSecondColumn verticalLR font24" data-offset-x="224" data-offset-y="171">A</span><span class="secondRowThirdColumn horizontalTB font12" data-offset-x="253" data-offset-y="344">A</span><span class="secondRowFourthColumn verticalLR font32" data-offset-x="221" data-offset-y="429">A</span>
        <span class="thirdRowFirstColumn  verticalLR font32" data-offset-x="119" data-offset-y="10">A</span><span class="thirdRowSecondColumn horizontalTB font12 extraLeftPadding" data-offset-x="137" data-offset-y="215">A</span><span class="thirdRowThirdColumn verticalLR font24" data-offset-x="122" data-offset-y="300">A</span><span class="thirdRowFourthColumn horizontalTB font16" data-offset-x="137" data-offset-y="482">A</span>
        <span class="fourthRowFirstColumn horizontalTB font24" data-offset-x="33" data-offset-y="51">A</span><span class="fourthRowSecondColumn verticalLR font32" data-offset-x="17" data-offset-y="155">A</span><span class="fourthRowThirdColumn horizontalTB font16 extraBottomPadding" data-offset-x="33" data-offset-y="337">A</span><span class="fourthRowFourthColumn verticalLR font12" data-offset-x="25" data-offset-y="469">A</span>
    </div>
offsetLeft expected 328 but got 342
FAIL .grid 5 assert_equals:
<div class="grid fit-content itemsBaseline verticalRL">
        <span class="firstRowFirstColumn verticalLR font12" data-offset-x="328" data-offset-y="73">A</span><span class="firstRowSecondColumn horizontalTB font16" data-offset-x="323" data-offset-y="170">A</span><span class="firstRowThirdColumn verticalLR font24 extraTopPadding" data-offset-x="323" data-offset-y="246">A</span><span class="firstRowFourthColumn horizontalTB font32 extraRightPadding" data-offset-x="323" data-offset-y="355">A</span>
        <span class="secondRowFirstColumn verticalLR font16" data-offset-x="221" data-offset-y="65">A</span><span class="secondRowSecondColumn horizontalTB font24 extraTopPadding" data-offset-x="221" data-offset-y="134">A</span><span class="secondRowThirdColumn verticalLR font12" data-offset-x="222" data-offset-y="294">A</span><span class="secondRowFourthColumn horizontalTB font32" data-offset-x="221" data-offset-y="355">A</span>
        <span class="thirdRowFirstColumn verticalLR font32 extraBottomPadding" data-offset-x="119" data-offset-y="10">A</span><span class="thirdRowSecondColumn horizontalTB font12" data-offset-x="151" data-offset-y="177">A</span><span class="thirdRowThirdColumn verticalLR font24" data-offset-x="122" data-offset-y="270">A</span><span class="thirdRowFourthColumn horizontalTB font16" data-offset-x="151" data-offset-y="380">A</span>
        <span class="fourthRowFirstColumn verticalLR font24 extraTopPadding" data-offset-x="17" data-offset-y="25">A</span><span class="fourthRowSecondColumn horizontalTB font32" data-offset-x="17" data-offset-y="145">A</span><span class="fourthRowThirdColumn verticalLR font16 extraBottomPadding" data-offset-x="21" data-offset-y="263">A</span><span class="fourthRowFourthColumn horizontalTB font12" data-offset-x="17" data-offset-y="387">A</span>
    </div>
offsetLeft expected 328 but got 342
PASS .grid 6
PASS .grid 7
PASS .grid 8
PASS .grid 9
PASS .grid 10
PASS .grid 11
PASS .grid 12
This test checks that baseline is applied correctly on a grid aling both axis and different writing-modes when using symbolic fonts, which will use middle-baseline.

A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A

A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A

A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A

A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A

A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A

A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A

A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A

A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A

A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A

A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A

A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A

A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
A
